Lower bit widths for label generation
Working with images in uint8 instead of float32 gives some major
benefits.

1. Lower memory usage. We can pack 4 pixels in to the space we used to
   pack one
2. Integer math is easier than floating point math

In our expensive sample - glyph diff, use integer math as much as
possible, and keep the bit widths down. This results in multiple inputs
being stored as uint8 instead of float32. Bring things back to float32
when needed

~40% speedup on CPU in visualize.py
~5% improvement on GPU during training, although unlocks a major
improvement down the line due to lower memory usage

+++ train.py
@@ -43,7 +43,7 @@ def render_full_image(img, glyph_cache, sample_width, sample_height, iter, net):
             num_y_samples * glyph_cache.shape[1],
             num_x_samples * glyph_cache.shape[2],
         ),
-        dtype=numpy.uint8,
+        dtype=numpy.float32,
     )
 
     char_idxs = torch.max(predictions, 1)[1]
